fix: avoid eslint errors
