+++ components/DoiContainer.tsx
@@ -4,8 +4,6 @@ import { gql } from "apollo-boost"
 import Error from "./Error"
 import { useQuery } from '@apollo/react-hooks';
 import Doi from './Doi'
-import { DoiType } from './types';
-
 
 type Props = {
   item?: string
@@ -69,6 +67,63 @@ export const DOI_GQL = gql`
   }
 }
 `
+
+interface DoiType {
+  id: string
+  doi: string
+  url: string
+  types: {
+    resourceTypeGeneral: string
+    resourceType: string
+  }
+  creators: Creator[]
+  titles: Title[]
+  publicationYear: number
+  publisher: string
+  descriptions: Description[]
+  rights: Rights[]
+  version: string
+  formattedCitation: string
+  citationCount: number
+  citationsOverTime: CitationsYear[]
+  citations: []
+  viewCount: number
+  viewsOverTime: UsageMonth[]
+  views: []
+  downloadCount: number
+  downloadsOverTime: UsageMonth[]
+  downloads: []
+}
+
+interface Creator {
+  id: string
+  name: string
+  givenName: string
+  familyName: string
+}
+
+interface Title {
+  title: string
+}
+
+interface Rights {
+  rights: string
+  rightsUri: string
+}
+
+interface Description {
+  description: string
+}
+
+interface CitationsYear {
+  year: string,
+  total: number
+}
+
+interface UsageMonth {
+  yearMonth: string,
+  total: number
+}
 
 
 interface DoiQueryData {
